+++ src/test.ts
@@ -1,3 +1,3 @@
-import { asp } from "./cli/cli";
+import { asp } from "./cli";
 
 asp(process.argv.slice(2));
